simpler gyroscope auto

+++ Team9450/src/main/java/org/firstinspires/ftc/team9450/subsystems/Drivetrain.java
@@ -78,7 +78,7 @@ public class Drivetrain extends Subsystem {
         setPower(new DriveSignal(power, power, power,power));
         try{while(isBusy());}catch (Exception e){}
     }
-    public void pivotTo(int pos, org.firstinspires.ftc.team9450.sensors.Gyroscope imu){
+    public void pivotTo(double pos, org.firstinspires.ftc.team9450.sensors.Gyroscope imu){
         while(imu.getAngle()!=pos){
             if(imu.getAngle()<pos){
                 setPower(new double[]{0.5,0.5,-0.5,-0.5});
